🚚 Moved and renamed `FAQTagList` component
