+++ src/components/FAQTagList.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { Link, StaticQuery, graphql } from 'gatsby'
-import { Spirit } from '../../spirit-styles'
+import { Spirit } from './spirit-styles'
 import PropTypes from 'prop-types'
-import { getTagsforPostCollection } from '../../../utils/tag-utils'
+import { getTagsforPostCollection } from '../utils/tag-utils'
 
 class FAQTagList extends React.Component {
     render() {
