+++ src/pages/faq.js
@@ -6,7 +6,7 @@ import Layout from '../components/layouts/default'
 import Link from 'gatsby-link'
 import { Spirit } from '../components/spirit-styles'
 import { MetaData } from '../components/meta'
-import FAQTagList from '../components/layouts/partials/faq-taglist'
+import FAQTagList from '../components/FAQTagList'
 import getMetaImageUrls from '../utils/getMetaImageUrls'
 
 class FAQPage extends React.Component {
